backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
python-multipart
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
